// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-impersonation-session-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminImpersonationSessionHttpResponse {
    actorEmail?: string | null;
    actorIpAddress?: string | null;
    actorUserAgent?: string | null;
    actorUserId?: string;
    id?: string;
    reason?: string;
    startedAtUtc?: string;
    targetEmail?: string | null;
    targetUserId?: string;
}