// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-user-creation-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminUserHttpResponse } from './admin-user-http-response';

export interface AdminUserCreationHttpResponse {
    credentialsEmailQueued?: boolean;
    temporaryPassword?: string;
    user?: AdminUserHttpResponse;
}