// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-user-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminUserHttpResponse {
    activityLevel?: string;
    aiConsentAcceptedAt?: string | null;
    aiInputTokenLimit?: number;
    aiOutputTokenLimit?: number;
    birthDate?: string | null;
    calorieCyclingEnabled?: boolean;
    carbTarget?: number | null;
    createdOnUtc?: string;
    dailyCalorieTarget?: number | null;
    dashboardLayoutJson?: string | null;
    deletedAt?: string | null;
    desiredWaistCm?: number | null;
    desiredWeightKg?: number | null;
    email?: string | null;
    fastingCheckInFollowUpReminderHours?: number;
    fastingCheckInReminderHours?: number;
    fastingPushNotificationsEnabled?: boolean;
    fatTarget?: number | null;
    fiberTarget?: number | null;
    firstName?: string | null;
    fridayCalories?: number | null;
    gender?: string | null;
    hasPassword?: boolean;
    heightCm?: number | null;
    hydrationGoal?: number | null;
    id?: string;
    isActive?: boolean;
    isEmailConfirmed?: boolean;
    language?: string | null;
    lastLoginAtUtc?: string | null;
    lastName?: string | null;
    latestWeightDate?: string | null;
    latestWeightKg?: number | null;
    mondayCalories?: number | null;
    mustChangePassword?: boolean;
    profileImage?: string | null;
    profileImageAssetId?: string | null;
    proteinTarget?: number | null;
    pushNotificationsEnabled?: boolean;
    roles?: Array<string>;
    saturdayCalories?: number | null;
    socialPushNotificationsEnabled?: boolean;
    stepGoal?: number | null;
    sundayCalories?: number | null;
    telegramUserId?: number | null;
    theme?: string | null;
    thursdayCalories?: number | null;
    tuesdayCalories?: number | null;
    uiStyle?: string | null;
    username?: string | null;
    waterGoal?: number | null;
    wednesdayCalories?: number | null;
    weightKg?: number | null;
}