// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-user-login-event-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminUserLoginEventHttpResponse {
    authProvider?: string;
    browserName?: string | null;
    browserVersion?: string | null;
    deviceType?: string | null;
    id?: string;
    loggedInAtUtc?: string;
    maskedIpAddress?: string | null;
    operatingSystem?: string | null;
    userAgent?: string | null;
    userEmail?: string | null;
    userId?: string;
}