// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-user-role-audit-event-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminUserRoleAuditEventHttpResponse {
    action?: string;
    actorEmail?: string | null;
    actorUserId?: string | null;
    id?: string;
    occurredAtUtc?: string;
    roleName?: string;
    source?: string;
    userId?: string;
}